DurationAdapter и LocalDateTimeAdapter перемещены в отдельный файл
