+++ src/main/java/ru/kermort/praktikum/taskmanager/DurationAdapter.java
@@ -1,0 +1,31 @@
+package ru.kermort.praktikum.taskmanager;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonToken;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.time.Duration;
+
+public class DurationAdapter extends TypeAdapter<Duration> {
+
+    @Override
+    public void write(final JsonWriter jsonWriter, final Duration duration) throws IOException {
+        if (duration == null) {
+            jsonWriter.nullValue();
+        } else {
+            jsonWriter.value(duration.toMinutes());
+        }
+    }
+
+    @Override
+    public Duration read(final JsonReader jsonReader) throws IOException {
+        if (jsonReader.peek() != JsonToken.NULL) {
+            return Duration.ofMinutes(jsonReader.nextLong());
+        } else {
+            jsonReader.nextNull();
+            return null;
+        }
+    }
+}
